migrate unit_states_test_kapacitor to pytest
